build: Add support for Android 15.
We try not to modify the AOSP repository, so the environment variable
and the ANDROID_12 macro do not exist, and an error will be reported
when the external repository references these header files.
However, new macros __ANDROID_VENDOR_API__ have been added in Android 15.

Type: Function
Redmine ID: #N/A
Associated modifications: N/A
Test: build A15

Signed-off-by: Bian Jin chen <kenjc.bian@rock-chips.com>
Change-Id: I9abb3a6f7d7ee20047c028a834c402086cf2c0e1