im2d_api: modify the default value of ‘interpolation’ in imresize()
Since the default value of 'interpolation' was defined as INTER_LINEAR
when imresize() was defined earlier, this parameter is actually not
effective. The RGA runs according to the default algorithm, so it is
more appropriate to configure it as IM_INTERP_DEFAULT here.

Also, since interpolation is supported, some default behaviors may
change to BI-Linear due to the INTER_LINEAR configuration, which is
incorrect.

Signed-off-by: Yu Qiaowei <cerf.yu@rock-chips.com>
Change-Id: I74551670384748e221ad70ecf688814eb91eabb6